Use the issue page of github; rm intern. page
The internationalization page seems to be defunct so I've removed it.

The "code" was coming out as red colour for whatever reason, which is not very useful in my opinion.

Also, we can just use the github issue page to list the issues, we
don't need our own page and error reporting form.