Remove links to defunct pages
Information and explanation about the layout pages and hidden pages is useful, but the links to specific pages on the website don't serve a useful purpose. The pages are already visible both in the listing and on the website menu already, and putting a list of pages here with explanations of what they are increases the amount of maintenance upon changing the pages, for no benefit I can think of.